chore: move controllers to separate directories
